add chrome specific features
